Fix: https://github.com/dterletskiy/carpc/issues/24
First we need to remove client pointer from the client map and only then process event
by client. The reason of this sequence is because during processing RESPONSE event
(inside the response) client could call the same related request, In this case client
will be added to the sequence-client map but will no be subscription on RESPONSE event,
because this will be second item in the map and according to logic (optimization)
subscription functionality will not be executed.
So now first we remove client from sequence-client map and only then process event by the client.
This was detected when client calles request_A, process response_A inside wich agine calles
request_A. In this case second response_A will not be processed.

Signed-off-by: Dmitriy Terletskiy <daterletskiy@gmail.com>